Move Transition and Delta matrix calculation into solver4x4 file
